libvncserver: scale: cast to 64 bit before shifting
Since pixel_value is 64 bit the data type of the shift operand should
be 64 bit too to prevent integer overflows.

+++ libvncserver/scale.c
@@ -210,7 +210,7 @@ void rfbScaledScreenUpdateRect(rfbScreenInfoPtr screen, rfbScreenInfoPtr ptr, in
              default:
                /* fixme: endianness problem? */
                for (z = 0; z < bytesPerPixel; z++)
-                 pixel_value += (srcptr2[z] << (8 * z));
+                 pixel_value += ((unsigned long)srcptr2[z] << (8 * z));
                 break;
               }
               /*
